img2txt/requirements.txt
Flask
torch
transformers
pillow
argostranslate
